+++ src/cli/commands/subnet/create.rs
@@ -75,7 +75,7 @@ pub struct CreateSubnetArgs {
     #[arg(long, short, help = "The name of the subnet")]
     pub name: String,
     #[arg(long, help = "The minimal validator stake amount (in whole FIL units)")]
-    pub min_validator_stake: u64,
+    pub min_validator_stake: f64,
     #[arg(long, help = "The minimal number of validators")]
     pub min_validators: u64,
     #[arg(long, help = "The bottom up checkpoint period in number of blocks")]
